Allow developers to edit files without restart
In order to test updates to modified .py files in Django, the Docker container
must be restarted. This complicates development and makes it hard to quickly
test updates. Add in configuration that causes the server to reload
automatically on file edit so testing can be expedited.

This update requires a rebuild of the docker container via
docker-compose up --build.

Signed-off-by: Frederick Lawler <fred@fredlawl.com>