Issue 66: docker dev env (#68)
I separated the services into 3 containers for ease of setup as well as easier logging. The services now simply log to stdout which can be viewed via docker logs and is the standard convention for logging in containers.

Also provided is a Makefile for easier to use aliases. make && make upd will build and start all services in detached mode. I kept running into a 502 error (connection error) when attempting to simply use the Django dev server w/ nginx as a reverse proxy so I went back to the route of using a file socket for comms. This is setup on a touch reload which can be activated with make reload.

I also disabled authentication when DEBUG=1 as once we start expanding the test suite we can avoid mocking the authentication in each request as well as make development easier.

Finally I made the index route basically a health monitor.

Resolves #66
Resolves #46
Resolves #47