Clean up model field imports (#101)
This PR explicitly imports the desired GIS fields as well as imports the base fields from Django models for a clearer reference to where the fields come from. Originally it appeared that `gis_models.CharField` and `base_models.CharField` are two different objects when in reality if the imports are followed the developer would find that they are the same.